début réfactoring
